backend/requirements.txt
# ── API server ────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Pinecone (current SDK — replaces pinecone-client) ──
pinecone>=3.2.2

# ── Ingestion pipeline ────────────────────────
pandas>=2.2.0
python-dotenv>=1.0.1

# ── Embeddings & Language ─────────────────────
sentence-transformers>=2.7.0
langdetect>=1.0.9
deep-translator>=1.11.4

# ── LLM providers ─────────────────────────────
google-generativeai>=0.5.4

# ── Validation & HTTP ─────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.1
httpx>=0.27.0
